Add a QRFactorization class.
Tasks 4 and 5 are about QR factorization. Methods that perform QR
factorization should be placed in this new class. For each QR method add
an entry in the qr_factorize_methods list, so that a wrapper is set up
in the Orthogonalize class and the method is added to the tests.